app/3_index_builder/requirements/common.txt
google-cloud-storage
google-cloud-logging
grpcio
langchain
pika
pypdf
sentence-transformers
cryptography>=3.1
llama_index==0.9.11
nest-asyncio
openai
llama-index
prisma==0.10.0